+++ components/calendar/src/vwc-calendar-event.ts
@@ -38,13 +38,6 @@ export class VWCCalendarEvent extends LitElement {
 	description?: string;
 
 	/**
-	 * @prop day - day of the week (starts from 0)
-	 * @public
-	 * */
-	@property({ type: Number, reflect: false })
-	day: 1 | 2 | 3 | 4 | 5 | 6 | 7 = 1;
-
-	/**
 	 * @prop color - color of event card
 	 * @public
 	 * */
@@ -80,15 +73,14 @@ export class VWCCalendarEvent extends LitElement {
 		const styles = {
 			...this.color && { '--vvd-calendar-event--primary-color': this.color },
 			...this.overlapCount && { '--vvd-calendar-event--overlap-count': this.overlapCount },
-			'--vvd-calendar-event--start': (this.start + 1).toString(),
+			'--vvd-calendar-event--start': (this.start).toString(),
 			'--vvd-calendar-event--duration': (this.duration).toString(),
-			'--vvd-calendar-event--day': this.day.toString()
 		};
 		return html`
 			<section
-				role="button"
-				tabindex="0"
-				style="${styleMap(styles)}"
+			role="button"
+			tabindex="0"
+			style="${styleMap(styles)}"
 			>
 				<h2><strong>${this.heading}</strong></h2>
 				<p>${this.description}</p>